refactor(ids): split four at-budget modules so the conjunction fix fits
The previous commit pushed `constraints/index.ts`, `parser/xml-parser.ts`,
`translation/service.ts` and `types.ts` past their module-size budgets,
each of which sat exactly at the file's line count. Rather than raise
them, cohesive blocks move out into new modules and the four rows
ratchet DOWN; `constraints/index.ts` drops under 400 and leaves the
allowlist entirely. No behaviour changes — every moved function is
byte-identical apart from de-indentation and the parameter that replaces
`this`.

  constraint-types.ts          the Constraint Types section of types.ts,
                               re-exported from it, so every existing
                               `from '../types.js'` import is unchanged
  parser/dom.ts                the DOM lookup helpers, which the parser
                               and its restriction parsing both need
  constraints/match-family.ts  the per-family matchers and their caches
  constraints/describe.ts      the expected-value display and the
                               mismatch reason
  translation/describe-constraint.ts
                               constraint wording and the interpolation
                               helper; the service delegates to both

`constraints/index.ts` is now the matching entry point alone. It keeps
re-exporting `formatConstraint` and `getConstraintMismatchReason`, so no
importer moves, and the dependency runs index -> match-family <-
describe, with no cycle.

The allowlist rows for `packages/cli/src/headless-backend.ts`,
`packages/mcp/src/backend-query.ts` and
`packages/merge/src/component-state.ts` were lowered by `--update` even
though this change does not touch those files; they are restored to
their recorded budgets and only the four ids rows move.

Refs #3272
