+++ packages/ids/src/constraints/index.ts
@@ -6,27 +6,9 @@
  * Constraint matching utilities for IDS validation
  */
 
-import type {
-  IDSConstraint,
-  IDSSimpleValue,
-  IDSPatternConstraint,
-  IDSEnumerationConstraint,
-  IDSBoundsConstraint,
-} from '../types.js';
+import type { IDSConstraint } from '../types.js';
 
-import {
-  compareBoolean,
-  compareNumeric,
-  compareString,
-  numericEpsilon,
-  isStrictNumericLiteral,
-  isBooleanLiteral,
-} from './comparators.js';
-import { isNumericXsdBase, isBooleanXsdBase } from './xsd-cast.js';
-import { translateXsdRegex } from './xsd-regex.js';
-
-/** Tolerance for the bounds matcher's exclusive comparators. */
-const NUMERIC_TOLERANCE = 1e-6;
+import { conjunctiveFacetsOf, matchOneFamily } from './match-family.js';
 
 /** Options for constraint matching */
 export interface MatchOptions {
@@ -68,483 +50,8 @@ export function matchConstraint(
   return true;
 }
 
-/**
- * The conjunctive facets `parseRestriction` attached to this constraint
- * from the same `<xs:restriction>`, or `undefined` when it declared one
- * facet family (the common case). `simpleValue` never carries any: it
- * does not come from a restriction.
- */
-function conjunctiveFacetsOf(
-  constraint: IDSConstraint
-): readonly IDSConstraint[] | undefined {
-  return constraint.type === 'simpleValue' ? undefined : constraint.and;
-}
 
-function matchOneFamily(
-  constraint: IDSConstraint,
-  actualValue: string | number | boolean,
-  ci: boolean
-): boolean {
-  switch (constraint.type) {
-    case 'simpleValue':
-      return matchSimpleValue(constraint, actualValue, ci);
-    case 'pattern':
-      return matchPattern(constraint, actualValue, ci);
-    case 'enumeration':
-      return matchEnumeration(constraint, actualValue, ci);
-    case 'bounds':
-      return matchBounds(constraint, actualValue);
-    default:
-      return false;
-  }
-}
-
-/**
- * Per-constraint comparator applicability. `compareNumeric` runs two
- * regex tests and `compareBoolean` two equality checks per call — and
- * simple-value matching sits inside per-entity × per-specification hot
- * loops (name matching against every pset/property). Whether the IDS
- * literal could EVER match numerically or boolean-ly depends only on
- * the constraint, so decide it once.
- */
-const SIMPLE_VALUE_COERCIBLE = new WeakMap<IDSSimpleValue, boolean>();
-
-function isCoercibleSimpleValue(constraint: IDSSimpleValue): boolean {
-  let coercible = SIMPLE_VALUE_COERCIBLE.get(constraint);
-  if (coercible === undefined) {
-    coercible =
-      isStrictNumericLiteral(constraint.value) ||
-      isBooleanLiteral(constraint.value);
-    SIMPLE_VALUE_COERCIBLE.set(constraint, coercible);
-  }
-  return coercible;
-}
-
-/**
- * Match against a simple value. Tries each comparator in order:
- * string → numeric → boolean. The first decisive result wins;
- * `undefined` lets the next strategy run.
- */
-function matchSimpleValue(
-  constraint: IDSSimpleValue,
-  actualValue: string | number | boolean,
-  caseInsensitive: boolean
-): boolean {
-  const expected = constraint.value;
-  const stringResult = compareString(expected, actualValue, caseInsensitive);
-  if (stringResult !== undefined) return stringResult;
-  // A non-numeric, non-boolean literal can only match through string
-  // equality — skip the comparators that would return undefined anyway.
-  if (!isCoercibleSimpleValue(constraint)) return false;
-  const numericResult = compareNumeric(expected, actualValue);
-  if (numericResult !== undefined) return numericResult;
-  const booleanResult = compareBoolean(expected, actualValue);
-  if (booleanResult !== undefined) return booleanResult;
-  return false;
-}
-
-/**
- * Match against a regex pattern
- * IDS uses XSD regex syntax which is slightly different from JavaScript
- */
-function matchPattern(
-  constraint: IDSPatternConstraint,
-  actualValue: string | number | boolean,
-  caseInsensitive = false
-): boolean {
-  // An XSD `xs:pattern` facet constrains the *lexical* space of its base
-  // datatype, so it matches the textual representation of the value — the
-  // official IDS reference (ifctester) does `re.fullmatch(pattern,
-  // str(value))`. A numeric value therefore satisfies e.g.
-  // `<restriction base="xs:decimal"><pattern value="^.*$"/>` ("any
-  // decimal value present"), which the previous blanket number/boolean
-  // bail-out wrongly failed on every numeric property.
-  //
-  // But the runtime value must first be type-compatible with the declared
-  // base: a number only matches a numeric base and a boolean only a
-  // boolean base. A number tested against `base="xs:string"` is a type
-  // mismatch — buildingSMART's corpus encodes exactly this as
-  // `patterns_always_fail_on_any_number`. (A string actual is already the
-  // lexical form, so it is matched directly regardless of base.)
-  if (typeof actualValue === 'number') {
-    if (!isNumericXsdBase(constraint.base)) return false;
-  } else if (typeof actualValue === 'boolean') {
-    if (!isBooleanXsdBase(constraint.base)) return false;
-  }
-  const actualStr = String(actualValue);
-
-  const regex = compilePatternRegex(constraint, caseInsensitive);
-  // An un-compilable pattern can't match anything (e.g. an unbalanced
-  // `[`); treat it as a non-match rather than throwing.
-  return regex ? regex.test(actualStr) : false;
-}
-
-/**
- * Compiled-pattern cache. Translating XSD → JS regex and building the
- * `RegExp` on every value check dominated pattern-heavy validation; the
- * compiled form depends only on (constraint, caseInsensitive), and the
- * constraint object is stable per parsed document.
- */
-const PATTERN_REGEX_CACHE = new WeakMap<
-  IDSPatternConstraint,
-  { cs?: RegExp | null; ci?: RegExp | null }
->();
-
-function compilePatternRegex(
-  constraint: IDSPatternConstraint,
-  caseInsensitive: boolean
-): RegExp | null {
-  let entry = PATTERN_REGEX_CACHE.get(constraint);
-  if (!entry) {
-    entry = {};
-    PATTERN_REGEX_CACHE.set(constraint, entry);
-  }
-  const slot = caseInsensitive ? 'ci' : 'cs';
-  let regex = entry[slot];
-  if (regex === undefined) {
-    regex = buildPatternRegex(constraint.pattern, caseInsensitive);
-    entry[slot] = regex;
-  }
-  return regex;
-}
-
-function buildPatternRegex(
-  xsdPattern: string,
-  caseInsensitive: boolean
-): RegExp | null {
-  // XSD char-class subtraction `[a-z-[aeiou]]` has no JS equivalent;
-  // approximate as the positive class (drop the exclusion) so the rest
-  // of the pattern still evaluates, matching long-standing behaviour.
-  const desubtracted = xsdPattern.replace(
-    /\[([^\]]+)-\[[^\]]+\]\]/g,
-    '[$1]'
-  );
-  // Shared XSD → JS translation: `\i`/`\c`/`\d`/`\w` (and their
-  // negations) map to Unicode property escapes, and verbatim `\p{…}`
-  // classes pass through — both require the `u` flag for full fidelity.
-  const { pattern } = translateXsdRegex(desubtracted);
-  // IDS patterns must match the entire lexical value. Wrapping in a
-  // non-capturing group anchors top-level alternation correctly
-  // (`a|b` → `^(?:a|b)$`, not `^a|b$`). Case-insensitive matching is
-  // opt-in per the call site (entity / predefined-type names use it;
-  // property and attribute values do not).
-  const anchored = `^(?:${pattern})$`;
-  try {
-    return new RegExp(anchored, caseInsensitive ? 'iu' : 'u');
-  } catch {
-    // Some patterns are valid under JS's lenient (Annex-B) dialect but
-    // rejected under `u`. Retry without it so plain patterns keep
-    // matching; this loses `\p{…}` fidelity for that one pattern only.
-    try {
-      return new RegExp(anchored, caseInsensitive ? 'i' : '');
-    } catch {
-      return null;
-    }
-  }
-}
-
-/**
- * Compiled exact-match sets per enumeration constraint. Real-world IDS
- * code lists carry hundreds of values and are matched against every
- * candidate entity, so the linear comparator walk dominated validation
- * time. Constraint objects are stable per parsed document, making a
- * WeakMap cache safe.
- */
-const ENUM_VALUE_SETS = new WeakMap<
-  IDSEnumerationConstraint,
-  { exact: Set<string>; upper: Set<string>; anyCoercible: boolean }
->();
-
-function getEnumValueSets(constraint: IDSEnumerationConstraint): {
-  exact: Set<string>;
-  upper: Set<string>;
-  anyCoercible: boolean;
-} {
-  let sets = ENUM_VALUE_SETS.get(constraint);
-  if (!sets) {
-    const exact = new Set(constraint.values);
-    const upper = new Set<string>();
-    let anyCoercible = false;
-    for (const v of constraint.values) {
-      upper.add(v.toUpperCase());
-      if (isStrictNumericLiteral(v) || isBooleanLiteral(v)) anyCoercible = true;
-    }
-    sets = { exact, upper, anyCoercible };
-    ENUM_VALUE_SETS.set(constraint, sets);
-  }
-  return sets;
-}
-
-/**
- * Match against an enumeration. The actual value matches if ANY of the
- * declared options matches under string / numeric / boolean comparison
- * — same strategy table as `matchSimpleValue`, just iterated.
- */
-function matchEnumeration(
-  constraint: IDSEnumerationConstraint,
-  actualValue: string | number | boolean,
-  caseInsensitive: boolean
-): boolean {
-  // O(1) fast path: a set hit is exactly the condition under which
-  // `compareString` would have returned true for some value, so this
-  // never changes the outcome — misses fall through to the full
-  // comparator walk for numeric / boolean semantics.
-  const sets = getEnumValueSets(constraint);
-  const actualStr = String(actualValue);
-  if (sets.exact.has(actualStr)) return true;
-  if (caseInsensitive && sets.upper.has(actualStr.toUpperCase())) return true;
-  // Pure-string enumerations are fully decided by the set lookups —
-  // only numeric/boolean literals can still match in the slow walk.
-  if (!sets.anyCoercible) return false;
-
-  return constraint.values.some((v) => {
-    const stringResult = compareString(v, actualValue, caseInsensitive);
-    if (stringResult !== undefined) return stringResult;
-    const numericResult = compareNumeric(v, actualValue);
-    if (numericResult !== undefined) return numericResult;
-    const booleanResult = compareBoolean(v, actualValue);
-    if (booleanResult !== undefined) return booleanResult;
-    return false;
-  });
-}
-
-/**
- * Match against numeric bounds
- */
-function matchBounds(
-  constraint: IDSBoundsConstraint,
-  actualValue: string | number | boolean
-): boolean {
-  // String-length facets (xs:length / xs:minLength / xs:maxLength)
-  // operate on the textual length, not on numeric magnitude. When any
-  // of them are present, evaluate the length constraints first.
-  if (
-    constraint.length !== undefined ||
-    constraint.minLength !== undefined ||
-    constraint.maxLength !== undefined
-  ) {
-    const str = String(actualValue);
-    if (constraint.length !== undefined && str.length !== constraint.length) {
-      return false;
-    }
-    if (constraint.minLength !== undefined && str.length < constraint.minLength) {
-      return false;
-    }
-    if (constraint.maxLength !== undefined && str.length > constraint.maxLength) {
-      return false;
-    }
-    // Length-only restrictions don't impose numeric bounds; if the
-    // constraint also carries min/max we fall through to the numeric
-    // check below (rare in practice).
-    if (
-      constraint.minInclusive === undefined &&
-      constraint.maxInclusive === undefined &&
-      constraint.minExclusive === undefined &&
-      constraint.maxExclusive === undefined
-    ) {
-      return true;
-    }
-  }
-
-  const num =
-    typeof actualValue === 'number'
-      ? actualValue
-      : parseFloat(String(actualValue));
-
-  if (isNaN(num)) return false;
-
-  if (
-    constraint.minInclusive !== undefined &&
-    num < constraint.minInclusive
-  ) {
-    return false;
-  }
-
-  if (
-    constraint.maxInclusive !== undefined &&
-    num > constraint.maxInclusive
-  ) {
-    return false;
-  }
-
-  if (constraint.minExclusive !== undefined && num <= constraint.minExclusive) {
-    return false;
-  }
-
-  if (constraint.maxExclusive !== undefined && num >= constraint.maxExclusive) {
-    return false;
-  }
-
-  return true;
-}
-
-/**
- * Cap enumeration rendering. These strings are embedded in per-entity
- * validation results — an uncapped 800-value code list produced ~20KB
- * per result and ballooned reports into the gigabytes (OOM crash on
- * large models). The full value list stays available on the constraint
- * object itself.
- */
-const MAX_ENUM_DISPLAY_VALUES = 10;
-
-function formatEnumValues(values: string[]): string {
-  const shown = values
-    .slice(0, MAX_ENUM_DISPLAY_VALUES)
-    .map((v) => `"${v}"`)
-    .join(', ');
-  const more = values.length - MAX_ENUM_DISPLAY_VALUES;
-  return more > 0 ? `[${shown}, … +${more} more]` : `[${shown}]`;
-}
-
-/**
- * Get a human-readable description of why a constraint match failed
- */
-export function getConstraintMismatchReason(
-  constraint: IDSConstraint,
-  actualValue: string | number | boolean | null | undefined
-): string {
-  if (actualValue === null || actualValue === undefined) {
-    return 'value is missing';
-  }
-
-  // Report the facet that actually rejected the value, not the primary
-  // one. With conjunctive facets the primary often matched and a
-  // sibling is what failed.
-  const siblings = conjunctiveFacetsOf(constraint);
-  if (siblings !== undefined && matchOneFamily(constraint, actualValue, false)) {
-    for (const sibling of siblings) {
-      if (!matchOneFamily(sibling, actualValue, false)) {
-        return getConstraintMismatchReason(sibling, actualValue);
-      }
-    }
-  }
-
-  switch (constraint.type) {
-    case 'simpleValue':
-      return `expected "${constraint.value}", got "${actualValue}"`;
-    case 'pattern':
-      return `"${actualValue}" does not match pattern "${constraint.pattern}"`;
-    case 'enumeration':
-      return `"${actualValue}" is not one of ${formatEnumValues(constraint.values)}`;
-    case 'bounds':
-      return getBoundsMismatchReason(constraint, actualValue);
-    default:
-      return 'unknown constraint type';
-  }
-}
-
-function getBoundsMismatchReason(
-  constraint: IDSBoundsConstraint,
-  actualValue: string | number | boolean
-): string {
-  const num =
-    typeof actualValue === 'number'
-      ? actualValue
-      : parseFloat(String(actualValue));
-
-  if (isNaN(num)) {
-    return `"${actualValue}" is not a valid number`;
-  }
-
-  const violations: string[] = [];
-
-  if (
-    constraint.minInclusive !== undefined &&
-    num < constraint.minInclusive - NUMERIC_TOLERANCE
-  ) {
-    violations.push(`must be >= ${constraint.minInclusive}`);
-  }
-
-  if (
-    constraint.maxInclusive !== undefined &&
-    num > constraint.maxInclusive + NUMERIC_TOLERANCE
-  ) {
-    violations.push(`must be <= ${constraint.maxInclusive}`);
-  }
-
-  if (constraint.minExclusive !== undefined && num <= constraint.minExclusive) {
-    violations.push(`must be > ${constraint.minExclusive}`);
-  }
-
-  if (constraint.maxExclusive !== undefined && num >= constraint.maxExclusive) {
-    violations.push(`must be < ${constraint.maxExclusive}`);
-  }
-
-  return `${num} ${violations.join(' and ')}`;
-}
-
-/**
- * Per-constraint display-string cache. Failure paths format the same
- * constraint for every non-matching entity — millions of times during
- * applicability filtering — and the output depends only on the
- * constraint object.
- */
-const FORMAT_CACHE = new WeakMap<IDSConstraint, string>();
-
-/**
- * Format a constraint for display
- */
-export function formatConstraint(constraint: IDSConstraint): string {
-  let formatted = FORMAT_CACHE.get(constraint);
-  if (formatted === undefined) {
-    formatted = formatConstraintUncached(constraint);
-    FORMAT_CACHE.set(constraint, formatted);
-  }
-  return formatted;
-}
-
-function formatConstraintUncached(constraint: IDSConstraint): string {
-  const own = formatOneFamily(constraint);
-  const siblings = conjunctiveFacetsOf(constraint);
-  if (siblings === undefined) return own;
-  // Conjunctive facets from the same restriction. Naming only the
-  // primary would report an expectation narrower than the one enforced.
-  return [own, ...siblings.map(formatOneFamily)].join(' and ');
-}
-
-function formatOneFamily(constraint: IDSConstraint): string {
-  switch (constraint.type) {
-    case 'simpleValue':
-      return `"${constraint.value}"`;
-    case 'pattern':
-      return `pattern "${constraint.pattern}"`;
-    case 'enumeration':
-      if (constraint.values.length === 1) {
-        return `"${constraint.values[0]}"`;
-      }
-      return `one of ${formatEnumValues(constraint.values)}`;
-    case 'bounds':
-      return formatBounds(constraint);
-    default:
-      return 'unknown';
-  }
-}
-
-function formatBounds(constraint: IDSBoundsConstraint): string {
-  const parts: string[] = [];
-
-  if (
-    constraint.minInclusive !== undefined &&
-    constraint.maxInclusive !== undefined
-  ) {
-    return `between ${constraint.minInclusive} and ${constraint.maxInclusive}`;
-  }
-
-  if (constraint.minInclusive !== undefined) {
-    parts.push(`>= ${constraint.minInclusive}`);
-  }
-
-  if (constraint.maxInclusive !== undefined) {
-    parts.push(`<= ${constraint.maxInclusive}`);
-  }
-
-  if (constraint.minExclusive !== undefined) {
-    parts.push(`> ${constraint.minExclusive}`);
-  }
-
-  if (constraint.maxExclusive !== undefined) {
-    parts.push(`< ${constraint.maxExclusive}`);
-  }
-
-  return parts.join(' and ') || 'any value';
-}
+export {
+  getConstraintMismatchReason,
+  formatConstraint,
+} from './describe.js';
